+++ src/main/java/com/chrisrm/idea/actions/themes/literature/club/MTSayoriThemeAction.java
@@ -24,26 +24,27 @@
  *
  */
 
-package com.chrisrm.idea.actions.themes;
+package com.chrisrm.idea.actions.themes.literature.club;
 
 import com.chrisrm.idea.MTConfig;
 import com.chrisrm.idea.MTThemeManager;
 import com.chrisrm.idea.MTThemes;
 import com.chrisrm.idea.actions.WeebShitManager;
-import com.chrisrm.idea.actions.accents.MTAmethystAccentAction;
+import com.chrisrm.idea.actions.accents.MTSkyAccentAction;
+import com.chrisrm.idea.actions.themes.MTAbstractThemeAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 
-public final class MTYuriThemeAction extends MTAbstractThemeAction {
-  private final MTAmethystAccentAction mtAmethystAccentAction = new MTAmethystAccentAction();
+public final class MTSayoriThemeAction extends MTAbstractThemeAction {
+  private final MTSkyAccentAction mtSkyAccentAction = new MTSkyAccentAction();
   @Override
   public boolean isSelected(final AnActionEvent e) {
-    return MTConfig.getInstance().getSelectedTheme() == MTThemes.YURI;
+    return MTConfig.getInstance().getSelectedTheme() == MTThemes.SAYORI;
   }
 
   @Override
   public void setSelected(final AnActionEvent e, final boolean state) {
-    mtAmethystAccentAction.actionPerformed(e);
-    MTThemeManager.getInstance().activate(MTThemes.YURI, true);
-    WeebShitManager.getInstance().activate(MTThemes.YURI);
+    mtSkyAccentAction.actionPerformed(e);
+    MTThemeManager.getInstance().activate(MTThemes.SAYORI, true);
+    WeebShitManager.getInstance().activate(MTThemes.SAYORI);
   }
 }
